Planning: used float in dp_poly_path and dp_st_graph.

+++ modules/planning/tasks/dp_poly_path/trajectory_cost.cc
@@ -52,7 +52,7 @@ TrajectoryCost::TrajectoryCost(
       vehicle_param_(vehicle_param),
       heuristic_speed_data_(heuristic_speed_data),
       init_sl_point_(init_sl_point) {
-  const double total_time =
+  const float total_time =
       std::min(heuristic_speed_data_.TotalTime(), FLAGS_prediction_total_time);
 
   num_of_time_stamps_ = static_cast<uint32_t>(
@@ -64,9 +64,9 @@ TrajectoryCost::TrajectoryCost(
     }
     auto sl_boundary = ptr_path_obstacle->PerceptionSLBoundary();
 
-    const double adc_left_l =
+    const float adc_left_l =
         init_sl_point_.l() + vehicle_param_.left_edge_to_center();
-    const double adc_right_l =
+    const float adc_right_l =
         init_sl_point_.l() - vehicle_param_.right_edge_to_center();
 
     if (adc_left_l + FLAGS_lateral_ignore_buffer < sl_boundary.start_l() ||
@@ -94,7 +94,7 @@ TrajectoryCost::TrajectoryCost(
             ptr_obstacle->GetPointAtTime(t * config.eval_time_interval());
 
         Box2d obstacle_box = ptr_obstacle->GetBoundingBox(trajectory_point);
-        constexpr double kBuff = 0.5;
+        constexpr float kBuff = 0.5;
         Box2d expanded_obstacle_box =
             Box2d(obstacle_box.center(), obstacle_box.heading(),
                   obstacle_box.length() + kBuff, obstacle_box.width() + kBuff);
@@ -106,24 +106,24 @@ TrajectoryCost::TrajectoryCost(
 }
 
 ComparableCost TrajectoryCost::CalculatePathCost(
-    const QuinticPolynomialCurve1d &curve, const double start_s,
-    const double end_s, const uint32_t curr_level, const uint32_t total_level) {
+    const QuinticPolynomialCurve1d &curve, const float start_s,
+    const float end_s, const uint32_t curr_level, const uint32_t total_level) {
   ComparableCost cost;
-  double path_cost = 0.0;
-  std::function<double(const double)> quasi_softmax = [this](const double x) {
-    const double l0 = this->config_.path_l_cost_param_l0();
-    const double b = this->config_.path_l_cost_param_b();
-    const double k = this->config_.path_l_cost_param_k();
+  float path_cost = 0.0;
+  std::function<float(const float)> quasi_softmax = [this](const float x) {
+    const float l0 = this->config_.path_l_cost_param_l0();
+    const float b = this->config_.path_l_cost_param_b();
+    const float k = this->config_.path_l_cost_param_k();
     return (b + std::exp(-k * (x - l0))) / (1.0 + std::exp(-k * (x - l0)));
   };
 
   const auto &vehicle_config =
       common::VehicleConfigHelper::instance()->GetConfig();
-  const double width = vehicle_config.vehicle_param().width();
+  const float width = vehicle_config.vehicle_param().width();
 
-  for (double curve_s = 0.0; curve_s < (end_s - start_s);
+  for (float curve_s = 0.0; curve_s < (end_s - start_s);
        curve_s += config_.path_resolution()) {
-    const double l = curve.Evaluate(0, curve_s);
+    const float l = curve.Evaluate(0, curve_s);
 
     path_cost += l * l * config_.path_l_cost() * quasi_softmax(std::fabs(l));
 
@@ -131,22 +131,22 @@ ComparableCost TrajectoryCost::CalculatePathCost(
     double right_width = 0.0;
     reference_line_->GetLaneWidth(curve_s + start_s, &left_width, &right_width);
 
-    constexpr double kBuff = 0.2;
+    constexpr float kBuff = 0.2;
     if (!is_change_lane_path_ && (l + width / 2.0 + kBuff > left_width ||
                                   l - width / 2.0 - kBuff < -right_width)) {
       cost.cost_items[ComparableCost::OUT_OF_BOUNDARY] = true;
     }
 
-    const double dl = std::fabs(curve.Evaluate(1, curve_s));
+    const float dl = std::fabs(curve.Evaluate(1, curve_s));
     path_cost += dl * dl * config_.path_dl_cost();
 
-    const double ddl = std::fabs(curve.Evaluate(2, curve_s));
+    const float ddl = std::fabs(curve.Evaluate(2, curve_s));
     path_cost += ddl * ddl * config_.path_ddl_cost();
   }
   path_cost *= config_.path_resolution();
 
   if (curr_level == total_level) {
-    const double end_l = curve.Evaluate(0, end_s - start_s);
+    const float end_l = curve.Evaluate(0, end_s - start_s);
     path_cost +=
         std::sqrt(end_l - init_sl_point_.l() / 2.0) * config_.path_end_l_cost();
   }
@@ -155,12 +155,12 @@ ComparableCost TrajectoryCost::CalculatePathCost(
 }
 
 ComparableCost TrajectoryCost::CalculateStaticObstacleCost(
-    const QuinticPolynomialCurve1d &curve, const double start_s,
-    const double end_s) {
+    const QuinticPolynomialCurve1d &curve, const float start_s,
+    const float end_s) {
   ComparableCost obstacle_cost;
-  for (double curr_s = start_s; curr_s <= end_s;
+  for (float curr_s = start_s; curr_s <= end_s;
        curr_s += config_.path_resolution()) {
-    const double curr_l = curve.Evaluate(0, curr_s - start_s);
+    const float curr_l = curve.Evaluate(0, curr_s - start_s);
     for (const auto &obs_sl_boundary : static_obstacle_sl_boundaries_) {
       obstacle_cost += GetCostFromObsSL(curr_s, curr_l, obs_sl_boundary);
     }
@@ -170,15 +170,15 @@ ComparableCost TrajectoryCost::CalculateStaticObstacleCost(
 }
 
 ComparableCost TrajectoryCost::CalculateDynamicObstacleCost(
-    const QuinticPolynomialCurve1d &curve, const double start_s,
-    const double end_s) const {
+    const QuinticPolynomialCurve1d &curve, const float start_s,
+    const float end_s) const {
   ComparableCost obstacle_cost;
-  double time_stamp = 0.0;
+  float time_stamp = 0.0;
   for (size_t index = 0; index < num_of_time_stamps_;
        ++index, time_stamp += config_.eval_time_interval()) {
     common::SpeedPoint speed_point;
     heuristic_speed_data_.EvaluateByTime(time_stamp, &speed_point);
-    double ref_s = speed_point.s() + init_sl_point_.s();
+    float ref_s = speed_point.s() + init_sl_point_.s();
     if (ref_s < start_s) {
       continue;
     }
@@ -186,9 +186,9 @@ ComparableCost TrajectoryCost::CalculateDynamicObstacleCost(
       break;
     }
 
-    const double s = ref_s - start_s;  // s on spline curve
-    const double l = curve.Evaluate(0, s);
-    const double dl = curve.Evaluate(1, s);
+    const float s = ref_s - start_s;  // s on spline curve
+    const float l = curve.Evaluate(0, s);
+    const float dl = curve.Evaluate(1, s);
 
     const common::SLPoint sl = common::util::MakeSLPoint(ref_s, l);
     const Box2d ego_box = GetBoxFromSLPoint(sl, dl);
@@ -197,23 +197,23 @@ ComparableCost TrajectoryCost::CalculateDynamicObstacleCost(
           GetCostBetweenObsBoxes(ego_box, obstacle_trajectory.at(index));
     }
   }
-  constexpr double kDynamicObsWeight = 1e-6;
+  constexpr float kDynamicObsWeight = 1e-6;
   obstacle_cost.safety_cost *=
       (config_.eval_time_interval() * kDynamicObsWeight);
   return obstacle_cost;
 }
 
 ComparableCost TrajectoryCost::GetCostFromObsSL(
-    const double adc_s, const double adc_l, const SLBoundary &obs_sl_boundary) {
+    const float adc_s, const float adc_l, const SLBoundary &obs_sl_boundary) {
   const auto &vehicle_param =
       common::VehicleConfigHelper::instance()->GetConfig().vehicle_param();
 
   ComparableCost obstacle_cost;
 
-  const double adc_front_s = adc_s + vehicle_param.front_edge_to_center();
-  const double adc_end_s = adc_s - vehicle_param.back_edge_to_center();
-  const double adc_left_l = adc_l + vehicle_param.left_edge_to_center();
-  const double adc_right_l = adc_l - vehicle_param.right_edge_to_center();
+  const float adc_front_s = adc_s + vehicle_param.front_edge_to_center();
+  const float adc_end_s = adc_s - vehicle_param.back_edge_to_center();
+  const float adc_left_l = adc_l + vehicle_param.left_edge_to_center();
+  const float adc_right_l = adc_l - vehicle_param.right_edge_to_center();
 
   if (adc_left_l + FLAGS_lateral_ignore_buffer < obs_sl_boundary.start_l() ||
       adc_right_l - FLAGS_lateral_ignore_buffer > obs_sl_boundary.end_l()) {
@@ -231,14 +231,14 @@ ComparableCost TrajectoryCost::GetCostFromObsSL(
     obstacle_cost.cost_items[ComparableCost::HAS_COLLISION] = true;
   }
 
-  const double delta_l = std::fabs(
+  const float delta_l = std::fabs(
       adc_l - (obs_sl_boundary.start_l() + obs_sl_boundary.end_l()) / 2.0);
 
   obstacle_cost.safety_cost +=
       config_.obstacle_collision_cost() *
       Sigmoid(config_.obstacle_collision_distance() - delta_l);
 
-  const double delta_s = std::fabs(
+  const float delta_s = std::fabs(
       adc_s - (obs_sl_boundary.start_s() + obs_sl_boundary.end_s()) / 2.0);
   obstacle_cost.safety_cost +=
       config_.obstacle_collision_cost() *
@@ -251,7 +251,7 @@ ComparableCost TrajectoryCost::GetCostBetweenObsBoxes(
     const Box2d &ego_box, const Box2d &obstacle_box) const {
   ComparableCost obstacle_cost;
 
-  const double distance = obstacle_box.DistanceTo(ego_box);
+  const float distance = obstacle_box.DistanceTo(ego_box);
   if (distance > config_.obstacle_ignore_distance()) {
     return obstacle_cost;
   }
@@ -266,15 +266,15 @@ ComparableCost TrajectoryCost::GetCostBetweenObsBoxes(
 }
 
 Box2d TrajectoryCost::GetBoxFromSLPoint(const common::SLPoint &sl,
-                                        const double dl) const {
+                                        const float dl) const {
   Vec2d xy_point;
   reference_line_->SLToXY(sl, &xy_point);
 
   ReferencePoint reference_point = reference_line_->GetReferencePoint(sl.s());
 
-  const double one_minus_kappa_r_d = 1 - reference_point.kappa() * sl.l();
-  const double delta_theta = std::atan2(dl, one_minus_kappa_r_d);
-  const double theta =
+  const float one_minus_kappa_r_d = 1 - reference_point.kappa() * sl.l();
+  const float delta_theta = std::atan2(dl, one_minus_kappa_r_d);
+  const float theta =
       common::math::NormalizeAngle(delta_theta + reference_point.heading());
   return Box2d(xy_point, theta, vehicle_param_.length(),
                vehicle_param_.width());
@@ -282,8 +282,7 @@ Box2d TrajectoryCost::GetBoxFromSLPoint(const common::SLPoint &sl,
 
 // TODO(All): optimize obstacle cost calculation time
 ComparableCost TrajectoryCost::Calculate(const QuinticPolynomialCurve1d &curve,
-                                         const double start_s,
-                                         const double end_s,
+                                         const float start_s, const float end_s,
                                          const uint32_t curr_level,
                                          const uint32_t total_level) {
   ComparableCost total_cost;
